fix exports issue with serverless
