+++ handler.js
@@ -10,9 +10,8 @@ const { promisify } = require('util');
  */
 const got = require('got');
 const Twit = require('twit');
-const Entities = require('html-entities').AllHtmlEntities;
 const { ungzip } = require('node-gzip');
-const entities = new Entities();
+const { getTitleAndDescription } = require('./helpers');
 
 /**
  * Environment handling
@@ -75,48 +74,6 @@ const getWebDocUrls = async () => {
   const webDocUrls = allDocUrls.filter(onlyAllowWebUrls);
 
   return webDocUrls;
-};
-
-/**
- * Read out <h1> and meta description for URL and check if the url holds a deprecated entry
- * We use the <h1> rather than the <title> as the title is a little more verbose
- *
- * @param {String} url
- * @returns {Promise} Array of h1 and description for the documented URL
- */
-const getTitleAndDescription = async (url) => {
-  const DESCRIPTION_REGEX = /<meta name="description" content="(.*?)"\/>/i;
-  const TITLE_REGEX = /<h1>(.*?)<\/h1>/i;
-  // to not rely on exact words this matches the deprecation container
-  const DEPRECATION_REGEX = /class="notecard deprecated"/;
-
-  const { body: doc } = await got(url);
-
-  if (DEPRECATION_REGEX.test(doc)) {
-    return [null, null];
-  }
-
-  const titleMatch = doc.match(TITLE_REGEX);
-
-  if (!titleMatch) {
-    return [null, null];
-  }
-
-  let [, title] = titleMatch;
-
-  if (title.length > 40) {
-    title = title.slice(0, 40) + '…';
-  }
-
-  const descriptionMatch = doc.match(DESCRIPTION_REGEX);
-
-  if (!descriptionMatch) {
-    return [null, null];
-  }
-
-  let [, description] = descriptionMatch;
-
-  return [entities.decode(title), entities.decode(description)];
 };
 
 /**
@@ -194,8 +151,4 @@ module.exports.tweet = async () => {
   } catch (e) {
     console.error(e);
   }
-};
-
-module.exports = {
-  getTitleAndDescription,
 };
